AMLS-4024 Refactored existing Supervision generators into reusable traits
